requirements.txt
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
ipython==7.21.0
isort==5.7.0
Markdown==3.3.4
pydantic==1.8.1
requests==2.25.1
